+++ Model/Product/Presenters/ProductPresenter.cs
@@ -4,7 +4,7 @@
     {
         public readonly Product Product;
 
-        public int Id { get { return Product.Id; } }
+        public long Id { get { return Product.Id; } }
         public string Name { get { return Product.Name; } }
         public string Category => Product.Category.FullPath;
         public string Manufacturer => Product.Manufacturer.Name;
